LayoutTests/imported/w3c/web-platform-tests/fetch/api/abort/general-serviceworker.https-expected.txt

CONSOLE MESSAGE: line 1714: Unhandled Promise Rejection: TypeError: null is not an object (evaluating 'remote.onerror = function(error) { this_obj.remote_error(error); }')

FAIL General fetch abort tests in a service worker null is not an object (evaluating 'remote.onerror = function(error) { this_obj.remote_error(error); }')
